refactor(api): extract shared corpus-index core from case-law and legislation twins
The case-law and legislation corpus-index handlers were near-identical
copies of one S3-chunked-load / sequential-ingest / compare-and-set audit
algorithm; every change had to land twice. Extract that orchestration into
lib/corpus-index/core.ts, parameterized over a per-family adapter that
supplies the domain surface (tables, batch queries, per-document shape,
audit-job writes).

Each handler keeps its route-facing exports (backfillCorpusIndex /
backfillLegislationCorpusIndex, removeDecisionFromCorpusIndex /
removeLegislationFromCorpusIndex, loadDocsForBatch) as thin re-exports of
the bound indexer, so callers need no changes. The per-process
ensuredIndexes cache stays one-per-family. The no-await-in-loop rationale
comments on the sequential ingest/CAS/delete paths move into the core,
once. Behavior is unchanged; the neutral entityId field on the shared
job-input type maps to each family's decisionId/documentId column in its
adapter.

Moving readCorpusText into the core also removes the legislation handler's
cross-domain import of case-law/corpus-storage.
