+++ apps/api/src/handlers/case-law/corpus-index.test.ts
@@ -63,7 +63,7 @@ describe("loadDocsForBatch read-failure isolation", () => {
     const failure = readFailures.at(0);
     expect(failure?.indexId).toBe(corpusIndexId(generation, "CZ"));
     expect(failure?.job).toMatchObject({
-      decisionId: badRow.id,
+      entityId: badRow.id,
       contentHash: "hash_bad",
       operation: "index",
       status: "failed",
